% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedConvenienceFunctions.R
\name{factorial}
\alias{factorial}
\title{Factorial}
\usage{
factorial(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value of factorial
}
\description{
Function to calculate the factorial of a variable
}
\examples{
factorial(5)

}
