% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedConvenienceFunctions.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value of geometric mean
}
\description{
Function to calculate the geometric mean of a variable
}
\examples{
x<-c(1,1,3,5,6,6)
gm_mean(x)

}
