% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedConvenienceFunctions.R
\name{Modes}
\alias{Modes}
\title{Mode}
\usage{
Modes(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of modes
}
\description{
Function to calculate the mode of a variable
}
\examples{
x<-c(1,1,3,5,6,6)
Modes(x)

}
