% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedConvenienceFunctions.R
\name{nonUnique}
\alias{nonUnique}
\title{Non-unique}
\usage{
nonUnique(x)
}
\arguments{
\item{x}{numeric or character vector}
}
\value{
numeric or character vector of non-unique values
}
\description{
Function that returns all non-unique values in a vector
}
\examples{
x<-c(1,1,3,5,6,6)
nonUnique(x)

}
